requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
joblib==1.3.2
linkedin_api @ git+https://github.com/tomquirk/linkedin-api.git@4d69958d1e31f9e5d962ef07712162e2dedaf440
lxml==5.1.0
numpy==1.26.3
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.2.0
